+++ rotkehlchen/chain/substrate/manager.py
@@ -433,7 +433,7 @@ class SubstrateManager():
                 type_registry_preset=si_attributes.type_registry_preset,
                 use_remote_preset=True,
             )
-        except (requests.exceptions.RequestException, WebSocketException) as e:
+        except (requests.exceptions.RequestException, WebSocketException, SubstrateRequestException) as e:  # noqa: E501
             message = (
                 f'{self.chain} could not connect to node at endpoint: {endpoint}. '
                 f'Connection error: {str(e)}.'