+++ rotkehlchen/tests/unit/test_assets.py
@@ -121,6 +121,10 @@ def test_cryptocompare_asset_support(cryptocompare):
         ethaddress_to_identifier('0x16980b3B4a3f9D89E33311B5aa8f80303E5ca4F8'),  # noqa: E501 # Kira Network (KEX) but another KEX in CC
         'DON',     # Donnie Finance but Donation Coin in CC
         'BAG',     # Baguette but not in CC
+        ethaddress_to_identifier('0xDe30da39c46104798bB5aA3fe8B9e0e1F348163F'),  # noqa: E501 # Gitcoin (GTC) but another GTC in CC
+        ethaddress_to_identifier('0x6D0F5149c502faf215C89ab306ec3E50b15e2892'),  # noqa: E501 # Portion (PRT) but another PRT in CC
+        'ANI',     # Animecoin (ANI) but another ANI in CC
+        'XEP',     # Electra Protocol (XEP) but another XEP in CC
     )
     for asset_data in GlobalDBHandler().get_all_asset_data(mapping=False):
         potential_support = (
@@ -261,6 +265,11 @@ def test_coingecko_identifiers_are_reachable():
         'XAI',  # Token suggestion doesn't match token in db
         'XPB',  # no Pebble Coin in Coingecko. Got other XPB symbol token
         'XNS',  # no Insolar in Coingecko. Got other XNS symbol token
+        'PIGGY',  # Coingecko listed another asset PIGGY that is not Piggy Coin
+        # coingecko listed CAR that is not our token CarBlock.io
+        ethaddress_to_identifier('0x4D9e23a3842fE7Eb7682B9725cF6c507C424A41B'),
+        # coingecko listed newb farm with symbol NEWB that is not our newb
+        ethaddress_to_identifier('0x5A63Eb358a751b76e58325eadD86c2473fC40e87'),
     )
     for asset_data in GlobalDBHandler().get_all_asset_data(mapping=False):
         identifier = asset_data.identifier